Revert "apkverify:Test apk digest computation"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/eac78595-4cce-4c46-a9f7-fcb5756191e3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I980c39d223f5b08268ced2de2662060fbd78b13c