Share a single VirtualMachine object for VMs
VirtualMachineManager.get() has always been returning a new
VirtualMachine object. This can cause confusion because VirtualMachine
object stores the VM's binder object which isn't shared among instances.
So it may be possible to run multiple guest VMs on a single instance,
which isn't advised.

This fixes it by maintaining a map from the VM to the corresponding
unique VirtualMachine object. VirtualMachineManager.get() will return an
existing object, rather than creating a new one. The mapping is removed
when the VM is deleted.

Bug: 238692795
Test: atest MicrodroidTests
Change-Id: Ie273e9118f4986aebf336b415d0e87a061228e1b