Revert "apkverify:Test apk digest computation"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/2612d613-5aa3-415d-a78b-c5b2a2e3cf3e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2f0edd5018a379fef85b280652c5f2d7407c090f