Revert "Re-connect after adb root"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/eef05304-3944-483a-8604-9e4fb94df356). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I85d1e7f4a3b9728b314bdafa6d672398e6f4169b