Merge logic of DICE HAL and diced in to dice-service
Remove a process from microdroid by having the diced and DICE HAL logic
in the same process. We don't need to worry about the vendor split in
microdroid so HALs aren't necessary.

The implementation of the diced interfaces is not complete as much of
what's there is overkill. It's likely this service will later migrate to
a custom interface that is more applicable to what's needed in
microdroid.

Bug: 243133253
Test: atest MicrodroidTests
Change-Id: I00e83f3afa96b9c49b8b76d6c695697e828af379