Revert "[apkverify] Merge idsig into apkverify package"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/bde62948-23ba-4508-9d8c-5fe199e84379). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1661170e933c24cb81fa157fc30eb0090dff23f1