Revert "[apkverify] Merge idsig into apkverify package"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/9e04c481-9175-413b-a029-228b114b6ed4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id1d0e7bf16b5b94e1d088b3116894817195fb1b2