Revert "[apkverify] Merge idsig into apkverify package"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/99d447e8-51a9-436d-97da-595e552a2e14). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I048ebf5d8175daa385af5932832b03f45b9950e3