Migrate VirtualMachine API to @SystemApi
Mark all the public members of the existing internal platform API as
@SystemApi.

Update system-current.txt to match.

Modify the demo app and the test APK and dependencies to build against
our stub library, instead of the implementation, and switch to the
system_current SDK rather than platform APIs. (We still use platform
APIs for the native payload code though.)

Bug: 243512107
Bug: 243512115
Test: atest MicrodroidTests MicrodroidHostTestCases
Test: Install & run demo app
CTS-Coverage-Bug: 244561836
Change-Id: Ibfffbb221f3ac37e03d5cfdbc9208428eeadbb9f