Measure app startup impact after VM teardown
Add a host side test as part of MicrodroidTestCase which starts a
Microdroid VM that reserves 80% of the device's memory reported from
/proc/meminfo (MemFree - field) and touches all the allocated guest memory
with a tmpfs filesystem and successive copies from /dev/zero.
This step is necessary to create host stage2 fragmentation and we
measure the cold app startup time after we tear-down the VM.

The following metrics will be reported from the adb shell of the Settings
app startup: TotalTime, WaitTime. This can be extended to add other apps
as well.

Bug: 236621823
Test: Verified localy atest MicrodroidTestCase
Signed-off-by: Sebastian Ene <sebastianene@google.com>
Change-Id: Iede9d6f0507c8d7e56b3e32324441ff11e54fa1e