Create IVirtualizationServiceInternal
Add a new AIDL interface for communication between per-app VS and
a singleton global VS. The global VS will remain in charge of allocating
globally-unique resources, such as the CID, and running singleton
servers, like tombstone receiver. The interface will never be exposed
outside of the split VS instances.

For now, create IVirtualizationServiceInternal and start a local
instance as part of the still unified VS process. This way we can start
moving functionality around before the final split.

As a first step, move CID allocation and tombstone server to the
internal service.

Bug: 245727626
Test: atest -p packages/modules/Virtualization:avf-presubmit
Change-Id: I43beac45bb6ed5b60182c62e8bb08d8d702192d0