Create IVirtualizationServiceInternal
Add new AIDL interface for communication between virtmgr and the global
virtualizationservice. This will be used for internal functionality,
synchronizing the instances of virtmgr, and never exposed to users.

Move the tombstone server to IVirtualizationServiceInternal. We will
eventually remove it completely, but for now there should only be one
instance of it.

Test: -
Change-Id: I43beac45bb6ed5b60182c62e8bb08d8d702192d0