Revert "Simplify use of config files"
Revert "Fix zipfuse race condition"

This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/f843a462-7d6e-4d88-baa7-3b58414486eb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib4f3a0a6c283e31d3657e00d6fb1ea55eb52613b