Revert "Report crosvm reboot reason on watchdog stall"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/2af3cd18-048d-4ff3-b2cb-414a05a5cc79). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I28d6bcdf8afc2057604bb4993a0d0a3ab8bb6f58