Revert "Report crosvm reboot reason on watchdog stall"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/3263c2da-e1fc-4a02-bc6c-67a8ce30daed). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6ccafd787a5437bdeebfa92f030a881e3e867a51