Revert "Add libopen_dice nostd wrapper and test it in vmbase_example"
Revert "Add libcrypto_baremetal"

This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/28b11a45-abde-4858-a87f-9a95ada2c20b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I61f392b04a024b7f930c109108aeb8c95cc6f363