Use ro.product.vendor.device to determine if running on Cuttlefish
So far, we used ro.product.name to check if the tests are running on
Cuttlefish or not. However, that is errorprone because Cuttlefish (as a
device) can run different variants of Android having different
ro.product.name values. For example, the prop is aosp_[arch] for GSI
(Generic System Image).

Fix this by using the device name that is in the vendor partition. The
partition is not replaced even with GSI.

Bug: 253638551
Test: run MicrodroidHostTestCases -- --abi x86_64 on Cuttlefish running
GSI

Change-Id: I6671d27a58bc4bcf7a209bbbfd7fdc5f0417ac68