pvmfw: Add support for appended configuration data
Implement a standardized way for pvmfw to receive data at load time,
which can be used by the platform to pass device-specific secrets or
influence the boot process. Unlike data received from the VMM, it is
(and must be) trusted.

Previously, the payload appended to pvmfw was the BCC (now incorporated
into the config data). To avoid breaking devices that do not yet support
this config format, if the appended data doesn't contain a valid config
header, fall back to assuming that it is a raw BCC when the 'legacy'
feature is enabled (default).

Bug: 238050226
Bug: 256827715
Test: atest MicrodroidTestApp
Change-Id: I2614e5df34df19052d7e12d24280d581dfaf06f7