pvmfw: Expect an appended BCC
In Android T, the BCC was passed by the platform to pvmfw by appending
it to its binary at load time. Expect this behavior and make use of the
BCC in main() by logging the slice, for now.

Implement the logic under a new 'legacy' module to isolate it as this
interface is expected to change for Android U.

Add the 'legacy' feature to control support for legacy interfaces; it is
intended to be disabled in the future.

Bug: 256827715
Test: atest MicrodroidTestApp
Change-Id: I4d0cfb028e42f2f7f5d343a941124f64d4d93f9a