Config file requires custom VM permission
Modify VS to require the USE_CUSTOM_VIRTUAL_MACHINE permission if a
config file is specified.

Most of the work here is in the tests. We request the permission in
the test APK. But we split test execution in two - most tests do not
have the custom permission granted (they shouldn't need it), while
those that do need it should have it granted.

Test: atest MicrodroidTestApp MicrodroidCustomConfigTestApp
Bug: 243513572
Change-Id: I0873e227c79c21de442e3c4158155dedbeb7f296