Revert "Move loopdevice module to dm_rust lib"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/bbb734db-19cd-4335-88fb-6bbccb9027ff). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I460a18cf487a6d22c33ffc8dcebc67cb095b9f81