Move android.system.virtualmachine jar to BCP
In Android U we are planning to provide privileged @SystemApi's to
manage virtual machines. These APIs will be provided by the
android.system.virtualmachine jar.

This change moves the android.system.virtualmachine jar to be part of
the BCP, which will allow us to use the processes and guidelines setup
for Android APIs.

Bug: 243512044
Test: build on product with AVF support
Test: build on product without AVF support
Test: atest --test-mapping packages/modules/Virtualization:avf-presubmit
Change-Id: Ifb0d04d58bb7e18fe548ab3780a0d753c8baba2d