pvmfw: smccc: Hack MMIO_GUARD_MAP to expect i32
As the kernel currently returns a signed 32-bit integer instead of the
SMCCC-compliant i64, perform the sign-extending cast in pvmfw. The bug
being fixed should not break the normal bootflow where the HVC succeeds,
as the HVC then returns 0.

Log the workaround, to help with debugging.

Note: this commit is intended to be reverted once the bug is fixed.

Bug: 253586500
Test: call mmio_guard_map() from non-protected VMs
Change-Id: I974e9783f114fb2d821b3806fdb69fb6c8eefdc7