pvmfw: Refactor MMIO_GUARD initialization
Refactor the MMIO_GUARD setup into its own module, add a clearer error
type allowing us to identify the faulting HVC in main_wrapper (e.g. for
logs - if they get through - or to map them to a future SYSTEM_RESET2
reboot reason) and let main() assume that logging is working.

This removes the only reason for main() to fail so temporarily remove
the Error type.

Test: m pvmfw_img && fastboot flash pvmfw && atest MicrodroidTestApp
Change-Id: I3dba728fee2c21e24195005e39593305aa3650f6