pvmfw: Hack MMIO_GUARD_UNMAP to expect NOT_SUPPORTED
As pKVM currently returns the wrong error code for the HVC (but
otherwise properly implements it), mask NOT_SUPPORTED in pvmfw. AFAICT,
as long as the bug is present, this error code will only every be
returned when the HVC succeeds so it should be pretty safe.

Keep returning Ok(()) on SUCCESS to avoid pvmfw breaking when the bug is
fixed.

Log the workaround, to help with debugging.

Note: this commit is intended to be reverted once the bug is fixed.

Bug: 251426790
Test: boot pvmfw with latest kernels
Change-Id: I74c2d43dd855e0d7f89fe286dea02eceea3c92be