Set RLIMIT_MEMLOCK of virtmgr, remove CAP_IPC_LOCK
Currently all instances of crosvm have CAP_IPC_LOCK, which allows them
to pin unlimited amount of memory. This is done by the kernel when
spawning a pVM, still in the context of the requesting process.

This does not work when spawning crosvm as a child (of a child) of the
calling process, as it cannot acquire more capabilities than its parent.
To work around this issue, give CAP_SYS_RESOURCE to the global
virtualizationservice and add a new method on
IVirtualizationServiceInternal that virtmgr can use to raise its (and
its children's) RLIMIT_MEMLOCK.

The value is currently set to 4GB but could be tweakable based on the
context of the caller, eg. their UID and permissions.

Test: -
Change-Id: I9b771d7fbd6376366ae19930bd93706a9b34b63e