Use per-VM temporary folder provided by the caller
VirtualizationService has a global folder for storing temporary files
while running VMs. These do not need to be privileged and currently do
not cound towards the storage used by the VM creator.

Instead of giving virtmgr permissions to write to this folder, have the
caller provide a temporary folder of their own. Since virtmgr runs with
the UID of the caller, this must be a folder that the caller can write
to.

Test: -
Change-Id: I3fb8338c5450a7e32261e74f0dcd2ee4cdd93570