Revert "Config file requires custom VM permission"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/894ad301-59c7-4e3c-b9d3-9e25cf2e98ad). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ide380e279dffc9aed9fd3a924937708ac0f1a196