Revert "Config file requires custom VM permission"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/df489448-de5b-45d6-8f74-1981f0f79a23). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0e8d96908347527d90e1ebc8dd86190e71d6e2a8