Join cgroup of the requesting process
Previously, when VM is started, the corresponding crosvm process joined
cgroup of its parent: virtualizationservice. This resulted in a problem
that the resource usage of the VM is not attributed to the owner of the
VM. The owner of the VM may abuse this to use VM as a means to use
system resource (cputime and memory) without being noticed and
controlled by the activity manager service (AMS).

Fixing this by putting the crosvm process to the same cgroup as its
owner. Then AMS will recognize the crosvm process as a phantom process
of the owning app and will adjust (e.g. kill when high memory pressure)
accordingly.

Bug: 245727626
Test: run a VM, check its cgroup by running
`cat /proc/$(pidof crosvm)/cgroup`

Change-Id: Ied1f646d58bbef6a13d94b4e2e82256c6f407643