Add cgroupof
cgroupof is a binary that returns cgroup of the given pid by reading
/proc/<pid>/cgroup. This is implemented as a separate binary instead of
as a library because reading the file requires some unusual privileges
and we often don't want to give the privileges to the callers.

1. AID_READPROC group. The /proc is mounted with `gid=3009 hidepid=2`
which means only the process in the group 3009 (AID_READPROC) can see
/proc/<pid> entries other than /proc/self.

2. read access to domain:file. /proc/<pid>/cgroup files are labeled as
the domain type of the owning process. In order for a process to read
"any" /proc/<pid>/cgroup, the process needs to have read access to
domain:file, which is too permissive for most processes.

With this implemented as a binary, we don't need to grant above
privileges to the callers. The privileges are granted only to this
binary, and the callers are only required to have access to execute this
binary.

Note that cgroupof has setgid (2xxx on the mode) on. See
canned_fs_config. Otherwise, it is executed as group of the caller and
hiddenpid=2 makes /proc/<pid> entries unavailable.

Bug: 245727626
Test: /apex/com.android.virt/bin/cgroupof $(pidof system_server)
returns uid_0/pid_929 as expected.

Change-Id: I0f41bce88c8338621af030296a3f3d76e471680e