Drop _debug from requester_debug_pid
The field was named as such because we allow the VM owner to exit while
the VM is still running. This is done by calling debugHoldVmRef to keep
the VM binder object in VM. As a result, the PID recorded in VS may
point to another process when queried after the original owner has died.

However, we don't use debugHoldVmRef for ordinary use cases including
our tests. Perhaps we may want to remove the method completely in the
future. Furthermore, the PID is guaranteed to be valid (i.e. points to
the owner/requester of VM) while we are in the process of creating the
VM. In order to avoid further confusion, drop _debug from the field
name.

Bug: N/A
Test: N/A
Change-Id: I7b6a5887c03c7e1e175509751b1558f24f470c9c