pvmfw: Fix the source of pvmfw_sign_key
pvmfw contains a reference public key against which it compares the key
used to sign the virtual machine payloads it receives. That reference is
built into pvmfw using the pvmfw_sign_key, a binary file containing the
key. For release builds, the pvmfw image is expected to be
post-processed to find-and-replace (in-place) that key with the public
release key.

Therefore, instead of incorrectly using the human-readable
avb_testkey_rsa4096 PEM file as pvmfw_sign_key, use the file containing
the public key in binary format, avb_testkey_rsa4096_pub_bin.

Bug: 255575669
Test: ?
Change-Id: Ia4490688cc3d1954a18e703a97812a9a28e63e7a