Enrich logs
The name of the VMs and identity of their requesters are shown in the
log. Previously, absence of this information made it difficult for
developers to identity which invocation of crosvm is for which
application.

Bug: 251751405
Test: run a VM and observe the log
Non-protected virtual machine "VmRunApp" (owner: "shell", cid: 10) created
Non-protected virtual machine "VmRunApp" (owner: "shell", cid: 10) started
Non-protected virtual machine "VmRunApp" (owner: "shell", cid: 10) stopped

Change-Id: I3678ca2257eb6437cb9ca7a7ad84643bb9612bc4