Run RpcBinder server for each VM on port==CID
Change CID allocation to allocate CIDs in the 0-65535 range (same as
ports), with the first 2048 reserved for platform use (such as port
2000 for tombstone receiver). Then run VirtualMachineService on the
same port as the VM's CID.

Change microdroid_manager to find the local CID and use it as port
number.

The host server will currently accept connections from all VMs. This
will be addressed in a follow-up CL.

Bug: 245727626
Test: atest -p packages/modules/Virtualization:avf-presubmit
Change-Id: Ia4c8448f76a0aa015658d6220c11fa73a8da8e73