Revert "microdroid_manager: print more info when payload fails v..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9830d9b9-6a31-405a-9208-7cfdd43d942d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iefaa05cb49859f0b2723df452526b7619cb5fdb1