Abstract get_calling_pid/uid based on context
When using kernel Binder, the caller can be identified using
IPCThreadState functions. This doesn't work in the context of RpcBinder
as the kernel does not provide that information.

Abstract the functions based on whether they are used by the global
virtualizationservice, which communicates over kernel Binder, or by
virtmgr which communicates over RpcBinder. In the latter case, the
values are:
  * calling_pid = parent PID
  * calling_uid = current UID

Bug: 245727626
Test: builds
Change-Id: If8c2fb3a28891186ef10e133971aebc81e091d23