Revert "Implement a memory reclaim microdroid benchmark test"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/87d9c714-afff-44e3-a9b9-bbacfbeac792). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7a85a9f851634c96211b859f63ddf71caefeaa41