Revert "Plumb onTrimMemory app callback to virtualizationservice"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/87d9c714-afff-44e3-a9b9-bbacfbeac792). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7f3e9d0bde21192eeb8ba10d1accac5cb99b536a