Revert "Move some VM config APIs to @TestApi"
Revert "Allow non-updatable modules to contribute to @TestApi"

This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/235ea371-8eb1-452b-9bed-355328f77f78). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie7730b3d85c1482243f214f8bae098b8298f20a4