Document ramdump pfd ownership
I added code to close the ParcelFileDescriptor when onRamdump()
returns, since I expect most apps won't do anything with it so it
seems unfair to put a burden on them. Documented this, and noted that
dup() can be used to bypass this if desired.

Also fixed a minor typo in passing.

Bug: 261037705
Test: Builds
Change-Id: I2d8d4ca09f71bb02a63a2754951516542042a786