Add libcap_rust
This is a rust library wrapping libcap C library.

We are only interested in being able to remove capabilities before
execve, hence the only public function libcap_rust has is
drop_caps(CapFlag : flag). Integration of this library with
microdroid_manager will be done in the follow-up change.

Additionally this change adds a basic test to ensure that drop_caps
doesn't fail.

Bug: 243633980
Test: atest libcap_rust.test
Change-Id: Ibfb45ec912df0ad0a1db62b24c22fbe5a61ff5f3