Revert "Expose the pvmfw module to the Make world"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d90823a0-c044-477e-ac7c-f702d4e0a2f3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I75b15b200796dfd1a07326e2e59fee67f04b5919