Revert "microdroid_manager: print more info when payload fails v..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bbe7082c-36a9-45c2-823e-ffa87eb1920b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I567c28538cbdbf00e2bbde5e43f5fc7b45cfc43f