Measure Microdroid boot time in non-debug mode
Our current benchmarks measure boot time in DEBUG_MODE_FULL and try to
parse the console output to identify time spent in individual boot
stages. That is helpful data but it means that the reported overall boot
time is not representative of production environment.

Rename the existing test to testMicrodroidBootTimeDebug and add
testMicrodroidBootTime that runs in DEBUG_MODE_NONE.

Test: atest MicrodroidBenchmarkApp
Change-Id: Iab706340803957ef26c6ec2cede75485c2e04808