Revert "[avb_footer] Include initrd_* hashes as hash descriptors..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f3e62710-ced5-4315-a205-4a42fd6c9a81). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If3f237148a49d46484a4409e1e6a9f0945b05f3e