Revert "[avb_footer] Include initrd_* hashes as hash descriptors..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/31ee73ef-dfcc-45e4-bb72-7eff62fab6de). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6d239f34e5a24bfbc9139f9027a0f97f03058ffa