Revert "Only accept binary name not path"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4a5e34fa-4145-4a68-a2a5-720a97ea3e1a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I16f5a73f744e48f180a906459394e1c44796ba1e