Revert "Only accept binary name not path"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/70b10378-1102-445b-85a0-5cca1774a91c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I74f62ca89375745f45c26c544e19637c5216241b