[split virtmgr] Separate code into own files
So as to avoid rebase issues during development, the separation between
virtmgr and virtualizationservice was done within one code base. Now
that the components are running separately at runtime, it is time to
separate them into two source folders as well.

As a first step, move code such that the two binaries do not share any
source files. These will be moved to the respective folders in a
subsequent change.

Bug: 245727626
Test: atest -p packages/modules/Virtualization:avf-presubmit
Change-Id: Ifcd7c4e32c36a0ba64ed1a955a289d6760e2010f