Mount /data with MS_NOEXEC
We want to ensure that W^X is guaranteed for Microdroid VMs. This change
doesn't guarantee that W^X is enforced for FULL_DEBUG VMs, as a user can
enable adb root, and remount the /data without MS_NOEXEC flag.

This is intended, as it allows developers to debug & get familiar with
the Microdroid execution environment.

Bug: 265261525
Test: atest MicrodroidTestApp
Change-Id: I622e3d95d9d8fd6d26bfb690499acf7208ca4d52