CTS to ensure images in VM's boot path use sha256
Android already has CTS to ensure this. Verified boot path in VM also
should use an algorithm at least as strong as sha256 & this should be
tested.

This patch covers this test. For this, we run avbtool info_image on the
images from virt apex & check the output. We do this again after
resigning the image. This ensures the CI covers the test for apex signed
using sign_virt_apex as well as by build infrastructure.

Bug: 263490360
Test: atest MicrodroidHostTests#testAllVbmetaUseSHA256
Change-Id: I92783d45f7841725ce20aa54f62767d2bbf6dd1d