[dice] Move hash function to the library libdiced_open_dice
And migrate the usage of the original function to the new one.

This is part of the project of merging the two existing dice
wrapper libraries into one library. The upstream library
libdiced_open_dice will be the merged library.

Test: atest diced_utils_test diced_sample_inputs_test \
diced_test diced_vendor_test diced_open_dice_cbor_test
Test: m pvmfw_img && m microdroid_manager
Test: atest microdroid_manager_test vmbase_example.integration_test
Bug: 267575445

Change-Id: Ic2c44b92dcc31bcc9c1fcd2317ae8058ef84faf6