Revert "Update source for Rust 1.66.1"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d7d5ff1d-5470-48e1-bc06-0b47195b4723). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If6d22107e1ea7f58fa607a2e593a37ce0082b230