Revert "virtualizationservice: Specify crashkernel only when nec..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3ebd8b90-367c-4756-8538-9b504fb45742). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icfb6d80aa0b383ddcb3b65afbce88b053b0f21f0