Set block size of encrypted storage to 4096B
If unspecified, block-size is heuristically determined by the filesystem
size and the expected usage of the filesystem. In case of encrypted
storage, the backing file(on host) anyway uses 4k, so the (default)
heuristic is not of much use. Increasing block size leads to better
performance of the used cipher (Adiantum), hence explicitly set the
block size to 4096.

Bug: 260329386
Test: stat -f /mnt/encryptedstore/
Change-Id: I87443bd1ee8cabf13585e7cbb5a3bbc4913a2d20