Revert "virtualizationservice: Specify crashkernel only when nec..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/781f46d4-e6d9-4c5b-abf8-c61e6a026419). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1e171417b2e00d376678a36b1d13d7b2870ac164