Revert "Temporally disable tombstone tests"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/781f46d4-e6d9-4c5b-abf8-c61e6a026419). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibf6650d06538616f7c18452f9da2107ebe9e18f8