microdroid_manager: Zero DICE hidden input
As we assume that instance.img entries can't be individually modified
without it being detected (thanks to the avf,new-instance flag), there
is no need to generate a random salt for DICE in microdroid_manager if
pvmfw has run (i.e. avf,strict-boot is received) so use a zero value if
one isn't found in the instance.img. This makes pvmfw the only source of
per-instance secret uniqueness, enabling end-to-end tests to validate
that functionality of pvmfw.

Bug: 268307476
Test: atest MicrodroidTestApp
Change-Id: Ia1c4f056b8106ddae4f2fd753fbc19e4c203b65a