Test that we can parcel & unparcel VirtualMachineDescriptor
The test creates a VirtualMachineDescriptor, writes it to a parcel,
reads it from the parcel, and then tries to create a VM from it.

This change also fixes the bug in how VirtualMachineDescriptor is
parceled/unparceled: the writeToParcel was writing a full
ParcelFileDescriptor parcelable, while VirtualMachineDescriptor(Parcel)
c-tor was reading just a raw fd.

There is also an interesting ownership issue with
VirtualMachineDescriptor. Right now the file descriptors in
VirtualMachineDescriptor are not duped, which means that if we send a
VirtualMachineDescriptor (say vmd_A) to the another app, the copy of the
VirtualMachineDescriptor (say vmd_B) will have the same fds. If after
that we somehow interact with vmd_A, it will affect the vmd_B as well.

I have left a TODO to either fix this by dupign the fds, or by
explicitly documenting this ownership semantics.

Bug: 259384440
Test: atest MicrodroidTestApp
Change-Id: I809fa635ebb934e0ba957d3e2642d1b40d21ec50