Make VM Descriptor AutoCloseable
I also close it implicitly on import - which seems natural, since we
do in fact close the PFDs as part of the import, so no further use can
be made of the descriptor.

Also disallow parcelling or importing once closed.

Add some related tests. (And, while I'm here, update the test
annotations - most of our tests aren't really checking that we can run
all available OSs in the VM.)

Bug: 268613460
Test: atest MicrodroidTests
Change-Id: I0558e4d0357faa5d113e6e1c2b2f8e59ea82c784