Don't start composd if VMs are not supported
If we attempt to it will abort (because virtmgr will abort), and
service manager will keep trying to sstart it forever.

To avoid having a third copy of the hypervisor properties code,
extract it to a library.

Fix "vm info" so it works even in the absence of VM support, by not
connecting to virtmgr when we don't need to.

Bug: 254599807
Test: "vm info", "vm list", "composd test-compile" with & without
    VM enabled
Change-Id: I1f33e231fcf9c77ce16a6b2cfb51d67da3986a6d