Revert "pvmfw: Remove allow(dead_code) on get_debug_policy"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/94e4e222-1588-4350-9ab0-ee0ba12481d6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I68da9f2ead8d7e4b948faa659f35b7482fe37c69