pvmfw: refactor: Centralize FDT modification
Instead of having unpack()/pack() pairs scattered around the codebase,
group all modifications of the DT at the end of main().

In particular, move the Debug Policy logic to the function, from
main_wrapper() where it used to be to please the borrow checker (now
fixed). This accommodates future extensions of the DP support and
addresses the confusing logging sequence we used to have e.g.

    [INFO] Starting payload...
    [ERROR] Unexpected error when handling debug policy: DebugPolicyFdt("Failed to apply overlay", NotFound)

Note: No functional change intended.

Bug: 249054080
Bug: 268265510
Test: atest MicrodroidHostTests
Change-Id: I1dc9d8107aa6e0f660d868914e0dfc4bab25ce42