Add IRemotelyProvisionedComponent skeleton impl inside RKP VM
This cl builds a VM and adds the skeleton code of
IRemotelyProvisionedComponent implementation.

We will turn this VM into RKP VM and integrate it into secure boot
in the future. The RKP VM will get attested against RKP server.

When a random VM wants to attest itself against RKP server, it is
supposed to communicate with RKP VM instead via VS. VS will then
launch the RKP VM and start the attestation service that will be
provided via RkpVm.

Test: Runs the RkpvmPayloadApk in VM
Bug: 241428822
Change-Id: Ib2bf573d232ecb6bf7bad86a15c7bcff0366dd24