[microdroid] Refactor and test format payload config descriptor
This cl extracted the DICE config descriptor formatting code into
a separate function and move it from main.rs to dice.rs.

We also added two tests for the newly created function. In a
subsequent cl, we are going to replace the manual cbor
serialization with the crate ciborium. The tests will ensure
that that this migration keep the formatting result unchanged.

Test: atest microdroid_manager_test
Test: Run microdroid
Bug: 268322533
Change-Id: If472431296ea24ac2917da023658677d2184655c