Shut down RpcServer together with crosvm
It can take a long time (maybe even never?) for Drop to be called on
a dead RpcBinder object. IVirtualMachine objects already rely on stop()
to be called when the client's VirtualMachine object is closed.
As a first step, shut down the vsock RpcServer running
VirtualMachineServer when the VM is killed.

Bug: 269731051
Test: atest MicrodroidTestApp
Change-Id: I6c9ca2f2d2dd424af206826db97e217e7f0ec5a5