[rialto] Enable Rialto running in protected VM
This allows Rialto to be used as the base of RKP VM
(WIP cl: aosp/2479977).

Prior to this cl, Rialto couldn't run in protected VM because MMIO
access is blocked by default in protected mode, this cl enables
the MMIO guard. We should also unmap the MMIO range used in Rialto
in the future.

Bug: 272226230
Test: atest rialto_test
Change-Id: I5307cb5d6cccc7a01af6bc9c46ae7111c0d51e93