Stop writing VM logs to files
The original reason was to allow us to direct potentially sensitive
logs to a well-protected location even on production devices.

But the current state is that the logs are only produced from debug
VMs, which we do in production. And it just means that when we have
test failures, we often don't see the logs that would tell us what is
going wrong.

Instead, just allow the logs to be sent to logcat. Get rid of test
code that attempted to preserve the now non-existent log files.

Bug: 275469579
Test: atest ComposBenchmarkApp; see VM logs in logcat
Test: composd test-compile; see VM logs in logcat
Change-Id: Ifa84b78b641a92b9f6a5290dfd6f3e169c20d53c