Revert "[pvmfw] Extract a library to support various hypervisor ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/efd9b024-08bf-4b43-ad31-05be5b15b076). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I94ee874b1ff0b40cbf9c3e31b034ddcf5ce70238