vmbase: Prevent aliasing of linker region markers
As we use the addr_of! macro, rustc can deduce that the end of a region
is different from the start of another because they are obtained from
the addresses of different variables. As those "variables" are simple
linker symbols, this assumption doesn't hold true and using it to
"optimize" the code (e.g. by removing `if` branches) leads to bugs.

Instead, compute the end address of the region through "start + length".

Test: atest vmbase_example.integration_test
Change-Id: I25ca7f9c42dd2da1af243f8fe41a1e07803a0d07