Update the getting started doc
This now covers the new Pixels (7 and 7 Pro). Sections that were needed
to recover the pvmfw partition are removed since 13 -> 14 upgrade path
doesn't have the issue. For the same reason, removed pixel6.md which was
about enabling AVF even before Android 13 is fully released. That doc
became obsolete with the public release of Android 13.

Bug: 278679146
Test: N/A

Change-Id: Ib35448039a2bdd8aad4fe8970cf77112ca64f184