Add BCC checking & truncating
Check whether any stage in the received BCC is marked as debug. If
not, refuse to apply any debug policy we receive. (The bootloader
shouldn't pass one in this case, this is just to make sure we catch
any mistake here.)

"Truncate" the received BCC by removing the entire chain we receive
and peforming a non-DICE derivation on the CDIs. This is to ensure
that we don't provide access to a UDS-rooted BCC, since that might be
what we received. This needs to be removed once we have a reliable way
to distinguish a VM BCC from a non-VM one.

In passing fix the lifetime of the config descriptor buffer
(b/280617929).

Fixed a test whose assumption is no longer true.

Bug: 266172411
Bug: 275424867
Test: atest ComposHostTestCases (this validates the CompOS BCC)
Test: atest MicrodroidTests
Change-Id: I507fedee9e21e8cbda60044a4e0324e0d6530b00

Add back the new truncation

Change-Id: Ifa1f8399e10c9bf214b0883ecbd495b4e70132b9