Add BCC checking & truncating
Extract BCC handling code in pvmfw to its own module, since we're
extending it.

Check whether any stage in the received BCC is marked as debug. If
not, refuse to apply any debug policy we receive. (The bootloader
shouldn't pass one in this case, this is just to make sure we catch
any mistake here.)

"Sanitise" the received BCC by removing everything before pvmfw from
it. This is to ensure that we don't provide access to a UDS-rooted
BCC, since that might be what we received. This needs to be removed
once we have a reliable way to distinguish a VM BCC from a non-VM one.

Bug: 266172411
Bug: 275424867
Test: atest ComposHostTestCases (this validates the CompOS BCC)
Test: atest MicrodroidTests
Change-Id: I507fedee9e21e8cbda60044a4e0324e0d6530b00