pvmfw: virtio: Stop zeroing allocated DMA regions
As the crate assumed that (parts of) the result of Hal::dma_alloc() was
zeroed and as alloc_shared() does not initialize the memory it allocates
(similarly to standard allocators e.g. GlobalAllocator), we had to zero
the region before returning to the virtio_drivers crate.

Now that the crate zeroes exactly what it needs to be (potentially less
than the whole region), there's no need to waste time doing in the HAL.
Note that the contents of regions allocated by alloc_shared() either
contain zeroes (following the very first allocation, as dynamically
shared pages are zeroed) or data that had previously been shared with
the host so this does not present a risk of leaking private (e.g.
heap-allocated) information.

Test: TH
Change-Id: I4e293439502140e0cd6e59f8b6e100e3bbe1c2c7