[WIP]vmbase: Implement time_this! for profiling
WIP: Re-implement this to not be so invasive on the source code e.g. by
using compiler hooks such as GCC's -finstrument-functions and/or Rust
proc macros.

Add a framework to record time elapsed by using the monotonic arch
counter and expose through a set of functions to read the clock and
record those times, and a convenience macro to wrap expressions.

Bug: 282936244
Test: atest ComposHostTestCases
Change-Id: Icc2b68fd4e8ee03a146a218b1a31671447c41006