Revert "[vbmeta] Raplace avb descriptor consts with AvbDescriptorTag"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9e28b784-8ae6-431c-9ac0-f36d1fc75ca4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I00aa2c705cc79e443753376fb0047e9463c7e75f