Revert "[vbmeta] Raplace avb descriptor consts with AvbDescriptorTag"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ecc75ddd-9479-4612-8c49-02d0a7c4957a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9458a08c22b9d92e072fcfdbab9493f73f26da3f