Revert "[avb] Refator descriptor module to allow multiple descri..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/930c7139-51ea-4564-ab54-3fd7346934b5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I47080874a07f0e31e9b62aabddf10ef2a9815691