Disable MicrodroidHostTests.testCustomVirtualMachinePermission.
Being able to revoke a granted signature permission is a bug even if the permission has the development flag, and the revocation will be lost upon any reconcilation of this permission, e.g. upon certain package changes or after system reboot. A correct way to test permission revocation can be using a test helper app that has a different signing certificate.

Bug: 288467613
Change-Id: I7eb65cdc52e195cc8317292dec9cf6284b0166f8
Test: MicrodroidHostTestCases