Monitor boottime metrics also from liblog logs
b/270566364 changed the Microdroid behavior so that most userspace logs
are emitted to liblog logs. Therefore, to measure the boottime metrics,
we need to read both console and liblog logs.b/270566364 changed the
Microdroid behavior so that most userspace logs are emitted to liblog
logs. Therefore, to measure the boottime metrics, we need to read both
console and liblog logs.

Bug: 270566364
Bug: 288849786
Test: atest MicrodroidBenchmarkApp
Change-Id: Ia2bff94acf53e51e66f56b00d396dec6e9735287