hyp: Introduce HypervisorCap::MMIO_GUARD
Similarly to DYNAMIC_MEM_SHARE, introduce a capability that client code
can check before calling MMIO_GUARD functions. This will remove the need
to implement stubs in the backends due to client code unconditionally
calling the functions.

Note that KvmHypervisor currently covers both pKVM and non-protected KVM
and will therefore have HypervisorCap::MMIO_GUARD, for now. As the only
client code that makes use of MMIO_GUARD and is intended to run on both
pKVM and regular KVM, Rialto, already deals with the distinction, this
patch doesn't deal with differentiating between the two modes, which is
done in a following patch.

No functional change intended.

Test: atest DebugPolicyHostTests#testNoAdbInDebugPolicy_withDebugLevelNone_boots
Test: atest rialto_test vmbase_example.integration_test
Change-Id: If45d31e65b2407deaf4df16cd4d354ccd6bf4506