vmbase: Initialize logger before calling main()
By setting up the console (MMIO_GUARD_MAP-ing it if necessary) and
initializing the logger before entering main, we can centralize the
early error handling instead of expecting clients to deal with it,
provide a standard interface for configuring the log level (relying on
the log crate) which can't fail, and simplify client code.

This will enable vmbase itself to log errors in future changes.

Test: atest MicrodroidTests#createAndConnectToVm
Test: atest rialto_test vmbase_example.integration_test
Change-Id: Id8c7256555d7940174c9f9375435d71851302020