Add a way to add a vendor disk image when launching Microdroid VM
This is a first patch in a series that adds support for "pluggable"
vendor images in Microdroid VMs.

In this patch virtmanager passes the vendor disk image to crosvm
invocation to make it available to guest Micrdroid.

Only callers with USE_CUSTOM_VIRTUAL_MACHINE permission can configure
the vendor image used in Microdroid.

Bug: 285855433
Test: presubmit
Test: vm run-microdroid --vendor /data/local/tmp/test_microdroid_vendor_image.img
Change-Id: I376468d6ef7dd675856f40b1361ca5d8a96cefb5