[vmbase] Move current sync exception handling to vmbase
The MemoryTracker in vmbase maps the MMIO memory range as lazy
device memory, which requires users of vmbase to handle
translation fault during the MMIO memory mapping. To avoid
duplicating the same fault handling code in each user, this cl
moves the fault handling to vmbase to streamline the process.

Bug: 284462758
Test: atest rialto_test vmbase_example.integration_test
Change-Id: Ie3c8f2d871b4ebf13763e36f952b5b71f4660e77